refactor(git-std): split cli/commit.rs into commit/mod.rs + prompt.rs + assemble.rs
